+++ src/App.js
@@ -1,5 +1,5 @@
 import Layout from "./components/layout/Layout";
-import Partner from "./pages/Partner";
+import Partner from "./pages/Partner/Partner";
 
 import './App.css';
 import './prismjs.css';
