+++ src/pages/Partner/Partner.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DataLine from "../../components/DataLine/DataLine";
 
-const Index = () => {
+const Partner = () => {
 	const categories = ['T1', 'T2', 'T3', 'T4', 'T5', 'T6', 'T7', 'T8', 'T9', 'T10', 'T11', 'T12'];
 	const series = [{
 		name: 'Doanh thu',
@@ -147,4 +147,4 @@ const Index = () => {
 	);
 };
 
-export default Index;
+export default Partner;
